Extract TransactionalGcsFileOutput as independent class file
