+++ src/test/java/org/embulk/output/TestGcsOutputPlugin.java
@@ -283,7 +283,7 @@ public class TestGcsOutputPlugin
     {
         ConfigSource configSource = config();
         PluginTask task = configSource.loadConfig(PluginTask.class);
-        Method method = GcsOutputPlugin.class.getDeclaredMethod("generateRemotePath", String.class, String.class, int.class, int.class, String.class);
+        Method method = GcsTransactionalFileOutput.class.getDeclaredMethod("generateRemotePath", String.class, String.class, int.class, int.class, String.class);
         method.setAccessible(true);
         assertEquals("sample.000.01.csv", method.invoke(plugin, "/sample", task.getSequenceFormat(), 0, 1, ".csv"));
         assertEquals("sample.000.01.csv", method.invoke(plugin, "./sample", task.getSequenceFormat(), 0, 1, ".csv"));
